refactoring all styling into one directory
