+++ spotify-analyzer/src/components/Page.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import '../App.css';
+import '../styles/App.css';
 
 import { authService } from '../services/auth';
