+++ spotify-analyzer/src/styles/App.css
