+++ spotify-analyzer/src/styles/BlurredPanel.css
