+++ spotify-analyzer/src/styles/Home.css
