+++ spotify-analyzer/src/styles/Settings.css
